Embeddings_VDB/Project 1 Enterprise Hybrid RAG System (Legal or Medical Domain)/requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
rank_bm25
python-dotenv
